requirements.txt
# Requirements for the project
# Requires Python 3.9

# PyTorch and CUDA
pytorch==1.12.1
torchvision==0.13.1
cudatoolkit==11.3

# Install PyG-related packages via: 
# pip install torch-scatter==2.1.0 torch-sparse==0.6.16 torch-cluster==1.6.0 torch-geometric==2.2.0 \
#   -f https://data.pyg.org/whl/torch-1.12.0+cu113.html
torch-scatter==2.1.0 
torch-sparse==0.6.16 
torch-cluster==1.6.0 
torch-geometric==2.2.0

# Other libraries
awkward==1.1.2
awkward-cpp==7
dgl==0.9.1
gdown==4.6.0
googledrivedownloader==0.4
h5py==3.7.0
joblib==1.2.0
json-tricks==3.16.1
matplotlib==3.6.3
munch==2.5.0
networkx==3.0
numpy==1.23.5
ogb==1.3.5
pandas==1.5.3
PyYAML==6.0
scikit-learn==1.2.0
scipy==1.10.0
simplejson==3.18.1
sparsebm==1.3
tqdm==4.64.1
uproot==4.0.4
optuna==4.1.0




